+++ Example/CZPicker/CZViewController.m
@@ -10,6 +10,8 @@
 
 @interface CZViewController ()
 @property NSArray *fruits;
+@property NSArray *fruitImages;
+@property CZPickerView *pickerWithImage;
 @end
 
 @implementation CZViewController
@@ -18,7 +20,9 @@
 {
     [super viewDidLoad];
     self.fruits = @[@"Apple", @"Banana", @"Grape", @"Watermelon", @"Lychee"];
+    self.fruitImages = @[[UIImage imageNamed:@"Apple"], [UIImage imageNamed:@"Banana"], [UIImage imageNamed:@"Grape"], [UIImage imageNamed:@"Watermelon"], [UIImage imageNamed:@"Lychee"]];
     // Do any additional setup after loading the view, typically from a nib.
+    self.title = @"CZPicker";
 }
 
 - (void)didReceiveMemoryWarning
@@ -46,23 +50,56 @@
     return self.fruits[row];
 }
 
-- (NSInteger)numberOfRowsInPickerView:(CZPickerView *)pickerView{
+- (UIImage *)czpickerView:(CZPickerView *)pickerView imageForRow:(NSInteger)row {
+    if([pickerView isEqual:self.pickerWithImage]) {
+        return self.fruitImages[row];
+    }
+    return nil;
+}
+
+- (NSInteger)numberOfRowsInPickerView:(CZPickerView *)pickerView {
     return self.fruits.count;
 }
 
-- (void)czpickerView:(CZPickerView *)pickerView didConfirmWithItemAtRow:(NSInteger)row{
+- (void)czpickerView:(CZPickerView *)pickerView didConfirmWithItemAtRow:(NSInteger)row {
     NSLog(@"%@ is chosen!", self.fruits[row]);
+    [self.navigationController setNavigationBarHidden:YES];
 }
 
--(void)czpickerView:(CZPickerView *)pickerView didConfirmWithItemsAtRows:(NSArray *)rows{
-    for(NSNumber *n in rows){
+- (void)czpickerView:(CZPickerView *)pickerView didConfirmWithItemsAtRows:(NSArray *)rows {
+    for (NSNumber *n in rows) {
         NSInteger row = [n integerValue];
         NSLog(@"%@ is chosen!", self.fruits[row]);
     }
 }
 
-- (void)czpickerViewDidClickCancelButton:(CZPickerView *)pickerView{
+- (void)czpickerViewDidClickCancelButton:(CZPickerView *)pickerView {
+    [self.navigationController setNavigationBarHidden:YES];
     NSLog(@"Canceled.");
+}
+
+- (void)czpickerViewWillDisplay:(CZPickerView *)pickerView {
+    NSLog(@"Picker will display.");
+}
+
+- (void)czpickerViewDidDisplay:(CZPickerView *)pickerView {
+    NSLog(@"Picker did display.");
+}
+
+- (void)czpickerViewWillDismiss:(CZPickerView *)pickerView {
+    NSLog(@"Picker will dismiss.");
+}
+
+- (void)czpickerViewDidDismiss:(CZPickerView *)pickerView {
+    NSLog(@"Picker did dismiss.");
+}
+
+- (IBAction)showWithImages:(id)sender {
+    self.pickerWithImage = [[CZPickerView alloc] initWithHeaderTitle:@"Fruits" cancelButtonTitle:@"Cancel" confirmButtonTitle:@"Confirm"];
+    self.pickerWithImage.delegate = self;
+    self.pickerWithImage.dataSource = self;
+    self.pickerWithImage.needFooterView = YES;
+    [self.pickerWithImage show];
 }
 
 - (IBAction)showWithFooter:(id)sender {
@@ -75,6 +112,7 @@
 
 - (IBAction)showWithoutFooter:(id)sender {
     CZPickerView *picker = [[CZPickerView alloc] initWithHeaderTitle:@"Fruits" cancelButtonTitle:@"Cancel" confirmButtonTitle:@"Confirm"];
+    picker.headerTitleFont = [UIFont systemFontOfSize: 40];
     picker.delegate = self;
     picker.dataSource = self;
     picker.needFooterView = NO;
@@ -87,5 +125,13 @@
     picker.dataSource = self;
     picker.allowMultipleSelection = YES;
     [picker show];
+}
+
+- (IBAction)showInsideContainer:(id)sender {
+    [self.navigationController setNavigationBarHidden:NO];
+    CZPickerView *picker = [[CZPickerView alloc] initWithHeaderTitle:@"Fruits" cancelButtonTitle:@"Cancel" confirmButtonTitle:@"Confirm"];
+    picker.delegate = self;
+    picker.dataSource = self;
+    [picker showInContainer:self.view];
 }
 @end